+++ tourcount/src/main/java/com/wmstein/tourcount/CountingLActivity.java
@@ -56,7 +56,7 @@ import androidx.core.app.NavUtils;
 import androidx.core.content.ContextCompat;
 
 /****************************************************************************************
- * CountingLActivity is necessary for overcoming some limitations of spinner.
+ * The 2. counting activity 'CountingLActivity' is necessary for overcoming some limitations of Android spinner.
  * It is the central activity of TourCount for landscape mode. 
  * It provides the same functionality as CountingActivity: the counters, 
  * starts GPS-location polling, starts EditIndividualActivity, 
@@ -68,7 +68,7 @@ import androidx.core.content.ContextCompat;
  *
  * Basic counting functions created by milo for BeeCount on 05/05/2014.
  * Adopted, modified and enhanced for TourCount by wmstein since 2016-04-18,
- * last modification on 2020-04-17
+ * last modification on 2020-04-23
  */
 public class CountingLActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener, PermissionsDialogFragment.PermissionsGrantedCallback
 {
@@ -94,6 +94,10 @@ public class CountingLActivity extends AppCompatActivity implements SharedPrefer
     // Location info handling
     private double latitude, longitude, height, uncertainty;
     LocationService locationService;
+
+    // Permission dispatcher mode modePerm: 
+    //  1 = use location service
+    //  2 = end location service
     int modePerm;
 
     private PowerManager.WakeLock mProximityWakeLock;